+++ src/part3-webserver/breakpoints/7-fibers/ws.ts
@@ -2,10 +2,14 @@ import {
   Chunk,
   Console,
   Effect,
+  Fiber,
+  FiberSet,
   HashMap,
   Layer,
   Match,
   Option,
+  PubSub,
+  Queue,
   Ref,
   Schedule,
   Stream,
@@ -38,11 +42,12 @@ const parseStartupMessage = pipe(S.parseJson(M.StartupMessage), S.decode);
 
 const initializeConnection = (
   ws: WebSocket,
-  publish: (message: M.ServerOutgoingMessage) => Effect.Effect<void>
+  sendToWsQueue: Queue.Dequeue<M.ServerOutgoingMessage>,
+  broadcastQueue: Queue.Enqueue<M.ServerOutgoingMessage>
 ) =>
   Effect.gen(function* (_) {
     console.log("Initializing connection");
-    const connectionsRef = yield* _(CurrentConnections);
+    const currentConnectionsRef = yield* _(CurrentConnections);
 
     const { color, name } = yield* _(
       Effect.async<M.StartupMessage, ParseError>((emit) => {
@@ -75,27 +80,7 @@ const initializeConnection = (
     );
 
     yield* _(Console.log(`New connection: ${name} (${color})`));
-    yield* _(publish({ _tag: "join", name, color }));
-
-    const connection: M.WebSocketConnection = {
-      _rawWS: ws,
-      name,
-      color,
-      timeConnected: Date.now(),
-      send: (message) =>
-        pipe(
-          message,
-          encodeMessage,
-          Effect.andThen((msg) => Effect.sync(() => ws.send(msg)))
-        ),
-      close: Effect.sync(() => ws.close()),
-    };
-
-    yield* _(
-      Ref.update(connectionsRef, (connections) =>
-        HashMap.set(connections, name, connection)
-      )
-    );
+    yield* _(Queue.offer(broadcastQueue, { _tag: "join", name, color }));
 
     const rawMessagesStream = Stream.async<string, Error>((emit) => {
       ws.on("message", (data) => {
@@ -114,7 +99,10 @@ const initializeConnection = (
       Stream.mapEffect(parseMessage),
       Stream.mapError(
         (parseError) => new M.UnknownIncomingMessageError({ parseError })
-      ),
+      )
+    );
+
+    const messagesWithInfoStream = parsedMessagesStream.pipe(
       Stream.map((message) => ({
         ...message,
         name,
@@ -125,44 +113,126 @@ const initializeConnection = (
       Stream.tapError((err) => Console.error(err))
     );
 
-    yield* _(Stream.runForEach(parsedMessagesStream, publish));
+    const broadcastFiber = yield* _(
+      Stream.runForEach(messagesWithInfoStream, (message) =>
+        Queue.offer(broadcastQueue, message)
+      ),
+      Effect.fork
+    );
+
+    const manualSendQueue = yield* _(
+      Queue.unbounded<M.ServerOutgoingMessage>()
+    );
+
+    const toSendStream = Stream.fromQueue(manualSendQueue).pipe(
+      Stream.merge(Stream.fromQueue(sendToWsQueue))
+    );
+
+    const sendToWsFiber = yield* _(
+      Stream.runForEach(toSendStream, (message) =>
+        encodeMessage(message).pipe(Effect.andThen((msg) => ws.send(msg)))
+      ),
+      Effect.fork
+    );
+
+    const connection: M.WebSocketConnection = {
+      _rawWS: ws,
+      name,
+      color,
+      timeConnected: Date.now(),
+      messages: parsedMessagesStream,
+      sendQueue: manualSendQueue,
+      close: Effect.sync(() => ws.close()),
+    };
+
+    yield* _(Effect.addFinalizer(() => connection.close));
+
+    yield* _(
+      Ref.update(currentConnectionsRef, (connections) =>
+        HashMap.set(connections, name, connection)
+      )
+    );
+
+    yield* _(Fiber.join(Fiber.zip(broadcastFiber, sendToWsFiber)));
   });
 
 export const Live = Layer.effectDiscard(
   Effect.gen(function* (_) {
     const wss = yield* _(WSSServer);
     const currentConnectionsRef = yield* _(CurrentConnections);
-    const publish = (message: M.ServerOutgoingMessage) =>
-      Effect.gen(function* (_) {
-        console.log("Publishing message", message);
-        const connections = yield* _(Ref.get(currentConnectionsRef));
-        yield* _(
-          Effect.forEach(HashMap.values(connections), (conn) =>
-            conn.send(message)
-          )
-        );
-      }).pipe(
-        Effect.catchAll((err) => Console.error(err)),
-        Effect.asUnit
-      );
 
-    const connectionsStream = createConnectionsStream(wss).pipe(
-      Stream.tapError((err) => Console.error(err))
+    const fiberSet = yield* _(
+      FiberSet.make<
+        void,
+        M.UnknownIncomingMessageError | M.BadStartupMessageError | ParseError
+      >()
     );
-    yield* _(
-      Stream.runForEach(connectionsStream, (ws) =>
-        initializeConnection(ws, publish)
-      )
+    const pubsub = yield* _(PubSub.unbounded<M.ServerOutgoingMessage>());
+
+    const connectionsStream = createConnectionsStream(wss);
+
+    const initializeConnectionsFiber = yield* _(
+      connectionsStream,
+      Stream.runForEach((ws) =>
+        Effect.gen(function* (_) {
+          const subscription = yield* _(pubsub.subscribe);
+          const fiber = yield* _(
+            initializeConnection(ws, subscription, pubsub),
+            Effect.fork
+          );
+          yield* _(FiberSet.add(fiberSet, fiber));
+        })
+      ),
+      Effect.fork
     );
-    yield* _(
+
+    const connectionLogFiber = yield* _(
       Effect.gen(function* (_) {
         const connections = yield* _(Ref.get(currentConnectionsRef));
         yield* _(
           Console.log(`Current connections: ${HashMap.size(connections)}`)
         );
+        for (const [name, connection] of HashMap.entries(connections)) {
+          yield* _(
+            Console.log(
+              `Connection: ${name} (${connection.color}) - ${Math.floor(
+                (Date.now() - connection.timeConnected) / 1000
+              )}s`
+            )
+          );
+        }
+
+        // forcefully sending a message to all clients
+        const message: M.ServerOutgoingMessage = {
+          _tag: "message",
+          name: "Server",
+          color: "white",
+          message: "THIS IS THE SERVER SPEAKING!",
+          timestamp: Date.now(),
+        };
+
+        // yield* _(Queue.offer(pubsub, message));
+
+        // just to one client
+        // forcefully sending a message to all clients
+        const ethanMessage: M.ServerOutgoingMessage = {
+          _tag: "message",
+          name: "Server",
+          color: "red",
+          message: "*i know its you ethan*",
+          timestamp: Date.now(),
+        };
+        const randomConnection = HashMap.get(connections, "ethan");
+        if (randomConnection._tag === "Some") {
+          yield* _(Queue.offer(randomConnection.value.sendQueue, ethanMessage));
+        }
       }),
-      Effect.repeat(Schedule.spaced("1 seconds"))
+      Effect.repeat(Schedule.spaced("1 seconds")),
+      Effect.fork
     );
-    // hmmmm why dont these work?
-  })
+
+    yield* _(
+      Fiber.join(Fiber.zip(initializeConnectionsFiber, connectionLogFiber))
+    );
+  }).pipe(Effect.forkDaemon)
 );